% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_hodgeslab_box.R
\name{geom_hodgeslab_box}
\alias{geom_hodgeslab_box}
\title{Geom for ggplot2 theme to add a box around the axes, for use in the Hodges lab}
\usage{
geom_hodgeslab_box(line_size = 0.5)
}
\arguments{
\item{line_size}{Sets the line width for lines used in plots, in units of pt. Defaults to 0.5.}
}
\description{
This function defines the box geom for plots coming from the Hodges lab.
}
\examples{
ggplot(df,aes(x=x,y=y)) + geom_point() + theme_hodgeslab_box() + geom_hodgeslab_box()
}
\keyword{box}
\keyword{hodgeslab}

