% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatScatterColor.R
\name{heatScatterColor}
\alias{heatScatterColor}
\title{Color values for x and y data using code from LSD::heatscatter(), from for the Hodges lab}
\usage{
heatScatterColor(
  x,
  y,
  coldiscrete = NULL,
  nrcol = 30,
  grid = 100,
  colpal = "heat",
  simulate = FALSE,
  daltonize = FALSE,
  cvd = "p",
  alpha = NULL,
  rev = FALSE,
  only = "none",
  nlevels = 10,
  greyscale = FALSE,
  log = "",
  ...
)
}
\arguments{
\item{x}{Values of x}

\item{y}{Values of y}

\item{colpal}{Color palette, options from LSD::disco(). Defaults to "heat".}

\item{...}{All other options the same as LSD::heatscatter()}
}
\description{
This takes in x, y, and returns color values based on a palette defined using the heatscatter() function from the "LSD" library.
Useful for adapting heatscatter() like behavior to plotting in ggplot2.
}
\examples{
heatScatterColor(x,y,coldiscrete,colpal="oranges")
}
\keyword{heatscatter}
\keyword{hodgeslab}
