% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatScatterDensityGroup.R
\name{heatScatterDensityGroup}
\alias{heatScatterDensityGroup}
\title{Discrete density group values for x and y data using code from LSD::heatscatter(), from for the Hodges lab}
\usage{
heatScatterDensityGroup(x, y, nrcol = 30, grid = 100, colpal = "heat",
  simulate = FALSE, daltonize = FALSE, cvd = "p", alpha = NULL,
  rev = FALSE, only = "none", nlevels = 10, greyscale = FALSE,
  log = "", ...)
}
\arguments{
\item{x}{Values of x}

\item{y}{Values of y}

\item{colpal}{Color palette, options from LSD::disco(). Defaults to "heat".}

\item{...}{All other options the same as LSD::heatscatter()}
}
\description{
This takes in x, y, and returns color values based on a palette defined using the heatscatter() function from the "LSD" library.
Useful for adapting heatscatter() like behavior to plotting in ggplot2.
}
\examples{
heatScatterDensityGroup(x,y,"oranges")
}
\keyword{heatscatter}
\keyword{hodgeslab}
