% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maplot.R
\name{maplot}
\alias{maplot}
\title{Plots MA plot for the Hodges lab}
\usage{
maplot(res, foldthresh = log2(1.5), pthresh = 0.1,
  colpal_unch = "greys", colpal_incr = "oranges",
  colpal_decr = "blues", minCount = 0, ylim = c(-2.25, 2.25),
  label = F, maxlabel = 50, grid = F, box = T, ...)
}
\arguments{
\item{res}{Results data frame from DESeq2. Important fields are "id", "baseMean", "log2FoldChange", "padj", and "label".}

\item{foldthresh}{Threshold for calling changes in either direction based on fold change. Defaults to log2(1.5).}

\item{pthresh}{Threshold for calling changes in either direction based on padj. Defaults to 0.10.}

\item{colpal_unch}{Color palette for unchanged sites, options from LSD::disco(). Defaults to "greys".}

\item{colpal_incr}{Color palette for increased sites, options from LSD::disco(). Defaults to "oranges".}

\item{colpal_decr}{Color palette for decreased sites, options from LSD::disco(). Defaults to "blues".}

\item{minCount}{Minimum x value to plot. Defaults to 0.}

\item{ylim}{Two-element vector setting the y-axis limits. Defaults to c(-2.25,2.25).}

\item{label}{Variable describing how or if labeling should occur. Can be TRUE, FALSE, or "auto". Defaults to FALSE.}

\item{maxlabel}{Indicates maximum number of labels to place on the plot. Only is evaluated when label == TRUE. Defaults to 50.}

\item{grid}{Boolean describing whether to show major gridlines. Defaults to FALSE.}

\item{box}{Boolean describing whether to place a box around the plot area. Defaults to TRUE.}
}
\description{
This function reads a data frame either directly from the results output of DESeq2, or by importing
a table saved directly from such a result data frame.
}
\examples{
maplot(res)
}
\keyword{hodgeslab}
\keyword{maplot}
