% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarySE.R
\name{summarySE}
\alias{summarySE}
\title{Summarizes data}
\usage{
summarySE(data = NULL, measurevar, groupvars = NULL, na.rm = FALSE,
  conf.interval = 0.95, .drop = TRUE)
}
\arguments{
\item{data}{A data frame.}

\item{measurevar}{The name of a column that contains the variable to be summariezed}

\item{groupvars}{A vector containing names of columns that contain grouping variables}

\item{na.rm}{A boolean that indicates whether to ignore NA's}

\item{conf.interval}{The percent range of the confidence interval (default is 95\%)}
}
\description{
Gives count, mean, standard deviation, standard error of the mean, and confidence interval (default 95\%).
Uses ddply() from the plyr library. From "Cookbook for R."
}
\examples{
df <- DNase; summarySE(df, measurevar="density", groupvars=c("Run","conc"))
df <- DNase; summarySE(df, measurevar="density", groupvars="conc")
}
\keyword{ci}
\keyword{hodgeslab}
\keyword{sd}
\keyword{sem}
\keyword{summarySE}
