% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_hodgeslab_basic.R
\name{theme_hodgeslab_basic}
\alias{theme_hodgeslab_basic}
\title{Basic ggplot2 theme for the Hodges lab}
\usage{
theme_hodgeslab_basic(
  base_size = 7,
  base_family = "",
  line_size = 0.25,
  grid = F,
  rotx = 0,
  box = F
)
}
\arguments{
\item{base_size}{Sets the Postscript font size for all labels, in units of pt. Defaults to 7.}

\item{base_family}{Sets the Postscript font family for all labels. Defaults to empty "".}

\item{line_size}{Sets the line width for lines used in plots, in units of pt. Defaults to 0.5.}
}
\description{
This function defines the basic theme for plots coming from the Hodges lab.
}
\examples{
ggplot(df,aes(x=x,y=y)) + geom_point() + theme_hodgeslab_basic()
}
\keyword{hodgeslab}
\keyword{theme}
